+++ api/instances/workbench-adapters.ts
@@ -1,5 +1,6 @@
-import { garbageLocalWorkbenchAdapter } from './garbage/workbench.js'
+import { localWorkbenchAdapterRegistration } from './garbage/workbench.js'
+import type { LocalWorkbenchAdapterRegistration } from '../workbench/local-adapter-types'
 
-export const registeredLocalWorkbenchAdapters = [
-  garbageLocalWorkbenchAdapter,
+export const registeredLocalWorkbenchAdapters: LocalWorkbenchAdapterRegistration[] = [
+  localWorkbenchAdapterRegistration,
 ]